Revert "Subscribe issue"
