+++ .gitignore
@@ -3,4 +3,3 @@ manifest.mf
 build.xml
 /target/
 andremury*
-paho*
