+++ src/main/java/com/mqtt/app/Publish/PublisherCallback.java
@@ -3,13 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.mqtt.app.subscribe;
+package com.mqtt.app.Publish;
 
-import com.mqtt.app.Config;
-import com.mqtt.app.controller.OperationController;
 import org.eclipse.paho.client.mqttv3.IMqttDeliveryToken;
 import org.eclipse.paho.client.mqttv3.MqttCallback;
-import org.eclipse.paho.client.mqttv3.MqttException;
 import org.eclipse.paho.client.mqttv3.MqttMessage;
 
 /**
@@ -18,7 +15,7 @@ import org.eclipse.paho.client.mqttv3.MqttMessage;
  *
  * @author andremury
  */
-public class SubscriberCallback implements MqttCallback {
+public class PublisherCallback implements MqttCallback {
 
     /**
      * This function is responsible to set a message when the server can't be
@@ -27,12 +24,7 @@ public class SubscriberCallback implements MqttCallback {
      * @param thrwbl Exception generated by the MQTT exception class.
      */
     public void connectionLost(Throwable thrwbl) {
-        System.out.println("Connection lost. Trying to reconnect..");
-        try {
-            reconnect();
-        } catch (MqttException ex) {
-            System.out.println("Connection has dropped. Please restart the subscriber module.");
-        }
+        System.out.println("Connection lost.");
     }
 
     /**
@@ -45,9 +37,8 @@ public class SubscriberCallback implements MqttCallback {
      */
     public void messageArrived(String string, MqttMessage mm) throws Exception {
         string = new String(mm.getPayload());
-        OperationController opc = new OperationController(string);
-        opc.init();
-
+        System.out.println("Reply: " + unjoin(string, ": ")[1] + ".");
+        ReplyGetter.unsubscribe();
     }
 
     /**
@@ -59,20 +50,7 @@ public class SubscriberCallback implements MqttCallback {
     public void deliveryComplete(IMqttDeliveryToken imdt) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
-
-    /**
-     * This function is responsible for the reconnect trying.
-     *
-     * @return Boolean success state.
-     * @throws MqttException
-     */
-    private boolean reconnect() throws MqttException {
-        try {
-            Subscriber.connect(Config.getTopic());
-            return true;
-        } catch (MqttException e) {
-            System.out.println(e);
-        }
-        return false;
+    private String[] unjoin(String payload, String delimiter) {
+        return payload.split(delimiter);
     }
 }
