+++ src/main/java/com/mqtt/app/Subscribe/Subscriber.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.mqtt.app.subscribe;
+package com.mqtt.app.Subscribe;
 
 import com.mqtt.app.Config;
 import org.eclipse.paho.client.mqttv3.MqttClient;
@@ -38,7 +38,6 @@ public class Subscriber {
     /**
      * Creates a connection.
      *
-     * @param topic
      * @return success state boolean
      * @throws MqttException
      */
@@ -59,11 +58,10 @@ public class Subscriber {
      * This function is responsible for disconnecting the subscriber from the
      * host.
      *
-     * @return Boolean success state.
+     * @return Bollean success state.
      */
     public boolean disconnect() {
         try {
-            client.unsubscribe(Config.getTopic());
             client.disconnect();
             return true;
         } catch (MqttException e) {
