+++ src/main/java/com/mqtt/app/Subscribe/SubscriberCallback.java
@@ -3,11 +3,13 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.mqtt.app.publish;
+package com.mqtt.app.Subscribe;
 
-import com.mqtt.app.services.ReplierService;
+import com.mqtt.app.Config;
+import com.mqtt.app.Controller.OperationController;
 import org.eclipse.paho.client.mqttv3.IMqttDeliveryToken;
 import org.eclipse.paho.client.mqttv3.MqttCallback;
+import org.eclipse.paho.client.mqttv3.MqttException;
 import org.eclipse.paho.client.mqttv3.MqttMessage;
 
 /**
@@ -16,7 +18,7 @@ import org.eclipse.paho.client.mqttv3.MqttMessage;
  *
  * @author andremury
  */
-public class PublisherCallback implements MqttCallback {
+public class SubscriberCallback implements MqttCallback {
 
     /**
      * This function is responsible to set a message when the server can't be
@@ -25,7 +27,12 @@ public class PublisherCallback implements MqttCallback {
      * @param thrwbl Exception generated by the MQTT exception class.
      */
     public void connectionLost(Throwable thrwbl) {
-        System.out.println("Connection lost.");
+        System.out.println("Connection lost. Trying to reconnect..");
+        try {
+            reconnect();
+        } catch (MqttException ex) {
+            System.out.println("Connection has dropped. Please restart the subscriber module.");
+        }
     }
 
     /**
@@ -38,9 +45,9 @@ public class PublisherCallback implements MqttCallback {
      */
     public void messageArrived(String string, MqttMessage mm) throws Exception {
         string = new String(mm.getPayload());
-        System.out.println("Reply: " + unjoin(string, ": ")[1] + ".");
-        ReplierService.setReply("Reply: " + unjoin(string, ": ")[1]);
-        ReplyGetter.unsubscribe();
+        OperationController opc = new OperationController(string);
+        opc.init();
+
     }
 
     /**
@@ -53,7 +60,19 @@ public class PublisherCallback implements MqttCallback {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
-    private String[] unjoin(String payload, String delimiter) {
-        return payload.split(delimiter);
+    /**
+     * This function is responsible for the reconnect trying.
+     *
+     * @return Boolean success state.
+     * @throws MqttException
+     */
+    private boolean reconnect() throws MqttException {
+        try {
+            Subscriber.connect(Config.getTopic());
+            return true;
+        } catch (MqttException e) {
+            System.out.println(e);
+        }
+        return false;
     }
 }
