+++ src/main/java/com/mqtt/app/states/Publisher.java
@@ -3,10 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.mqtt.app.models;
+package com.mqtt.app.states;
 
 import com.mqtt.app.Config;
-import com.mqtt.app.publish.ReplyGetter;
+import com.mqtt.app.Publish.ReplyGetter;
 import org.eclipse.paho.client.mqttv3.MqttClient;
 import org.eclipse.paho.client.mqttv3.MqttException;
 import org.eclipse.paho.client.mqttv3.MqttMessage;
@@ -32,7 +32,7 @@ public class Publisher {
     /**
      * T@var clientId This is the client ID tha will be generated later.
      */
-    private static String clientId;
+    private String clientId;
 
     /**
      * This is the class constructor that will initialize every dependency and
@@ -57,7 +57,7 @@ public class Publisher {
      */
     public boolean publish(String message, String topic) throws MqttException {
         try {
-            msg.setPayload((clientId + ": " + message).getBytes());
+            msg.setPayload((this.clientId + ": " + message).getBytes());
             client.publish(topic, this.msg);
             return true;
         } catch (MqttException e) {
@@ -92,6 +92,7 @@ public class Publisher {
     public boolean disconnect() throws MqttException {
         try {
             this.client.disconnect();
+            System.out.println("Client disconnected.");
             return true;
         } catch (MqttException e) {
             System.out.println(e);
@@ -109,7 +110,10 @@ public class Publisher {
      */
     public static boolean getReply() throws MqttException, InterruptedException {
         ReplyGetter psub = new ReplyGetter();
-        return psub.connect(clientId);
+        if (psub.connect("andremury")) {
+            return true;
+        }
+        return false;
     }
 
 }
